refactor(crawler): extract URL normalization and utilities from discovery.rs
- Extract binary file helpers to infrastructure/crawler/binary_utils.rs
- Extract robots.txt enforcement to infrastructure/crawler/robots_utils.rs
- Update discovery.rs to use extracted modules
- Update engine.rs and scrape_flow.rs to import from infrastructure layer
- All 1552 tests pass, clippy clean
