+++ src/application/crawler/discovery.rs
@@ -10,6 +10,7 @@ use url::Url;
 use crate::application::url_filter::is_allowed;
 use crate::domain::{CrawlError, CrawlerConfig, DiscoveredUrl, ScrapedContent, ValidUrl};
 use crate::error::{Result as ScraperResult, ScraperError};
+use crate::infrastructure::crawler::binary_utils::derive_filename_from_response;
 use crate::infrastructure::crawler::{
     extract_links, is_internal_link, normalize_url, SitemapConfig, SitemapParser,
 };
@@ -20,253 +21,6 @@ use crate::ScraperConfig;
 use crate::infrastructure::observability::metrics_instruments::{
     CRAWLER_BANDWIDTH, CRAWLER_PAGES, CRAWLER_URLS,
 };
-
-use std::sync::Arc;
-
-use dashmap::DashMap;
-use robotstxt::DefaultMatcher;
-
-// ============================================================================
-// Binary file download helpers
-// ============================================================================
-
-/// Simple percent-decoding for filenames (handles common cases).
-fn percent_decode(input: &str) -> String {
-    let mut result = String::with_capacity(input.len());
-    let mut chars = input.chars();
-    while let Some(c) = chars.next() {
-        if c == '%' {
-            let hex: String = chars.by_ref().take(2).collect();
-            if let Ok(byte) = u8::from_str_radix(&hex, 16) {
-                result.push(byte as char);
-            } else {
-                result.push('%');
-                result.push_str(&hex);
-            }
-        } else {
-            result.push(c);
-        }
-    }
-    result
-}
-
-/// Derive a filename from Content-Disposition header or URL path.
-///
-/// Priority: Content-Disposition `filename` > URL path basename > fallback.
-fn derive_filename_from_response(
-    headers: &wreq::header::HeaderMap,
-    url: &Url,
-    content_type: &str,
-) -> String {
-    // Try Content-Disposition header first
-    if let Some(disposition) = headers.get(wreq::header::CONTENT_DISPOSITION) {
-        if let Ok(val) = disposition.to_str() {
-            // Parse filename*=UTF-8''encoded or filename="name"
-            if let Some(name) = parse_content_disposition(val) {
-                return name;
-            }
-        }
-    }
-
-    // Derive from URL path
-    let path = url.path();
-    let basename = path.rsplit('/').next().unwrap_or("");
-    if !basename.is_empty() && basename != "/" {
-        // Clean up the basename — remove query params that may be appended
-        let clean = basename.split('?').next().unwrap_or(basename);
-        if !clean.is_empty() {
-            return clean.to_string();
-        }
-    }
-
-    // Fallback: generate filename from content type
-    let ext = match content_type {
-        ct if ct.contains("application/pdf") => "pdf",
-        ct if ct.contains("application/zip") => "zip",
-        ct if ct.contains("application/x-tar") => "tar",
-        ct if ct.contains("image/png") => "png",
-        ct if ct.contains("image/jpeg") => "jpg",
-        ct if ct.contains("image/gif") => "gif",
-        ct if ct.contains("image/webp") => "webp",
-        ct if ct.contains("image/svg") => "svg",
-        ct if ct.contains("audio/mpeg") => "mp3",
-        ct if ct.contains("video/mp4") => "mp4",
-        _ => "bin",
-    };
-
-    // Use URL host + path hash for uniqueness
-    let host = url.host_str().unwrap_or("unknown");
-    let path_hash = {
-        use std::collections::hash_map::DefaultHasher;
-        use std::hash::{Hash, Hasher};
-        let mut hasher = DefaultHasher::new();
-        path.hash(&mut hasher);
-        format!("{:x}", hasher.finish())
-    };
-    format!("{}_{}.{ext}", host.replace('.', "_"), &path_hash[..8])
-}
-
-/// Parse Content-Disposition header value to extract filename.
-///
-/// Supports:
-/// - `filename="report.pdf"`
-/// - `filename=report.pdf`
-/// - `filename*=UTF-8''encoded-name.pdf`
-fn parse_content_disposition(value: &str) -> Option<String> {
-    // Try filename*= first (RFC 5987 encoding)
-    for part in value.split(';') {
-        let part = part.trim();
-        if let Some(rest) = part.strip_prefix("filename*=") {
-            // Format: UTF-8''encoded_name
-            if let Some(name) = rest.strip_prefix("UTF-8''") {
-                // Simple percent-decoding for common cases
-                let decoded = percent_decode(name);
-                if !decoded.is_empty() {
-                    return Some(decoded);
-                }
-            }
-        }
-    }
-
-    // Try filename= (standard)
-    for part in value.split(';') {
-        let part = part.trim();
-        if let Some(rest) = part.strip_prefix("filename=") {
-            let name = rest.trim_matches(|c| c == '"' || c == '\'');
-            if !name.is_empty() {
-                return Some(name.to_string());
-            }
-        }
-    }
-
-    None
-}
-
-// ============================================================================
-// Robots.txt Enforcement
-//
-// Functions in this section will be wired into the Engine in PR #6.
-// ============================================================================
-
-/// Parsed robots.txt rules for a domain.
-///
-/// Following **api-non-exhaustive**: can add fields without breaking changes.
-/// Following **own-arc-shared**: wrapped in `Arc` for cache sharing.
-#[derive(Debug, Clone)]
-pub(crate) struct RobotsRules {
-    /// Raw robots.txt content for the robotstxt matcher.
-    content: String,
-    /// Parsed Crawl-delay in seconds, if present.
-    #[allow(dead_code)]
-    crawl_delay_secs: Option<f64>,
-}
-
-/// Cache of robots.txt rules keyed by domain.
-///
-/// Using `DashMap` for lock-free concurrent reads during crawl.
-/// No TTL — robots.txt rarely changes during a single crawl session.
-pub(crate) type RobotsCache = DashMap<String, Arc<RobotsRules>>;
-
-/// Create a new empty robots.txt cache.
-#[must_use]
-pub(crate) fn new_robots_cache() -> RobotsCache {
-    DashMap::new()
-}
-
-/// Parse Crawl-delay from raw robots.txt content.
-///
-/// Searches for `Crawl-delay:` directives (case-insensitive) and returns
-/// the first valid numeric value found.
-fn parse_crawl_delay(content: &str) -> Option<f64> {
-    for line in content.lines() {
-        let trimmed = line.trim();
-        if trimmed.to_lowercase().starts_with("crawl-delay:") {
-            if let Some(val_str) = trimmed.split(':').nth(1) {
-                if let Ok(val) = val_str.trim().parse::<f64>() {
-                    return Some(val);
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Fetch and cache robots.txt rules for a domain.
-///
-/// On cache miss, fetches `robots.txt` from the domain root using wreq.
-/// Parses the content and caches the result. Returns `None` if fetching
-/// or parsing fails (fail-open: treat as all-allowed).
-async fn fetch_robots_rules(domain: &str, cache: &RobotsCache) -> Option<Arc<RobotsRules>> {
-    if let Some(rules) = cache.get(domain) {
-        return Some(Arc::clone(rules.value()));
-    }
-
-    let robots_url = format!("https://{domain}/robots.txt");
-    tracing::debug!("Fetching robots.txt from {}", robots_url);
-
-    let content = match wreq::get(&robots_url).send().await {
-        Ok(resp) if resp.status().is_success() => match resp.text().await {
-            Ok(text) => text,
-            Err(e) => {
-                tracing::warn!("Failed to read robots.txt body for {}: {}", domain, e);
-                return None;
-            },
-        },
-        Ok(resp) => {
-            tracing::debug!(
-                "robots.txt for {} returned status {}, treating as all-allowed",
-                domain,
-                resp.status()
-            );
-            return None;
-        },
-        Err(e) => {
-            tracing::warn!("Failed to fetch robots.txt for {}: {}", domain, e);
-            return None;
-        },
-    };
-
-    let crawl_delay = parse_crawl_delay(&content);
-    let rules = Arc::new(RobotsRules {
-        content,
-        crawl_delay_secs: crawl_delay,
-    });
-    cache.insert(domain.to_string(), Arc::clone(&rules));
-    Some(rules)
-}
-
-/// Check if a URL is allowed by the site's robots.txt.
-///
-/// Fetches robots.txt on first encounter (cached per domain).
-/// Uses the `robotstxt` crate's `DefaultMatcher` for path matching.
-/// Fail-open: if robots.txt cannot be fetched, the URL is allowed.
-///
-/// # Arguments
-///
-/// * `url` - The full URL to check
-/// * `domain` - The domain key for cache lookup
-/// * `cache` - Shared robots.txt rules cache
-///
-/// # Returns
-///
-/// `true` if the URL is allowed by robots.txt (or if robots.txt is unavailable).
-pub(crate) async fn is_allowed_by_robots(url: &str, domain: &str, cache: &RobotsCache) -> bool {
-    let rules = match fetch_robots_rules(domain, cache).await {
-        Some(r) => r,
-        None => return true, // fail-open
-    };
-
-    let mut matcher = DefaultMatcher::default();
-    matcher.one_agent_allowed_by_robots(&rules.content, "*", url)
-}
-
-/// Get the crawl-delay for a domain in seconds, if configured.
-///
-/// Returns `None` if no Crawl-delay directive was found.
-#[allow(dead_code)]
-pub(crate) fn get_crawl_delay(domain: &str, cache: &RobotsCache) -> Option<f64> {
-    cache.get(domain).and_then(|r| r.crawl_delay_secs)
-}
 
 // ============================================================================
 // TUI Support — Discover/Scrape Use Cases
@@ -557,7 +311,7 @@ pub async fn scrape_single_url_for_tui(
                         );
                     } else {
                         info!(
-                            "💾 Saved binary file: {} ({} bytes)",
+                            "Saved binary file: {} ({} bytes)",
                             output_path.display(),
                             bytes.len()
                         );
@@ -648,7 +402,7 @@ pub async fn scrape_single_url_for_tui(
             })
         },
         Err(e) => {
-            warn!("⚠️  Readability failed for {}: {}", url, e);
+            warn!("Readability failed for {}: {}", url, e);
             let fallback_content = fallback::extract_text(&extraction_html);
 
             // Check if fallback produced poor content (likely extraction failure)
@@ -660,7 +414,7 @@ pub async fn scrape_single_url_for_tui(
                     MIN_FALLBACK_CONTENT,
                     e
                 );
-                warn!("⚠️  {}", msg);
+                warn!("{}", msg);
                 return Err(ScraperError::ExtractionFailed {
                     url: url.to_string(),
                     reason: msg,
@@ -1180,293 +934,5 @@ mod tests {
         assert_eq!(urls.len(), 2);
         assert!(urls.contains(&"https://example.com/page1".to_string()));
         assert!(urls.contains(&"https://external.com/page2".to_string()));
-    }
-
-    #[test]
-    fn test_robots_txt_disallow_enforcement() {
-        let robots_body = "\
-User-agent: *
-Disallow: /private/
-Disallow: /admin/
-Allow: /public/
-
-User-agent: BadBot
-Disallow: /";
-
-        // URL under Disallow should be blocked
-        let mut matcher = DefaultMatcher::default();
-        assert!(
-            !matcher.one_agent_allowed_by_robots(
-                robots_body,
-                "*",
-                "https://example.com/private/secret"
-            ),
-            "/private/ should be disallowed"
-        );
-        assert!(
-            !matcher.one_agent_allowed_by_robots(
-                robots_body,
-                "*",
-                "https://example.com/admin/config"
-            ),
-            "/admin/ should be disallowed"
-        );
-
-        // URL under Allow should be permitted
-        assert!(
-            matcher.one_agent_allowed_by_robots(
-                robots_body,
-                "*",
-                "https://example.com/public/page"
-            ),
-            "/public/ should be allowed"
-        );
-
-        // Unrestricted URL should be permitted
-        assert!(
-            matcher.one_agent_allowed_by_robots(robots_body, "*", "https://example.com/other"),
-            "/other should be allowed (no rule)"
-        );
-    }
-
-    #[test]
-    fn test_robots_txt_crawl_delay_parsing() {
-        let robots_body = "\
-User-agent: *
-Crawl-delay: 5
-Disallow: /tmp/";
-
-        assert_eq!(parse_crawl_delay(robots_body), Some(5.0));
-
-        let no_delay = "User-agent: *\nDisallow: /";
-        assert_eq!(parse_crawl_delay(no_delay), None);
-
-        let fractional = "User-agent: *\nCrawl-delay: 0.5\n";
-        assert_eq!(parse_crawl_delay(fractional), Some(0.5));
-    }
-
-    #[test]
-    fn test_robots_txt_crawl_delay_case_insensitive() {
-        let robots_body = "user-agent: *\nCrawl-Delay: 10\n";
-        assert_eq!(parse_crawl_delay(robots_body), Some(10.0));
-
-        let robots_body_upper = "User-Agent: *\nCRAWL-DELAY: 3\n";
-        assert_eq!(parse_crawl_delay(robots_body_upper), Some(3.0));
-    }
-
-    #[tokio::test]
-    async fn test_robots_cache_hit() {
-        let cache = new_robots_cache();
-        let rules = Arc::new(RobotsRules {
-            content: "User-agent: *\nDisallow: /private/\n".to_string(),
-            crawl_delay_secs: Some(2.0),
-        });
-        cache.insert("example.com".to_string(), rules);
-
-        // Should allow public URL
-        assert!(is_allowed_by_robots("https://example.com/public", "example.com", &cache).await);
-        // Should disallow private URL
-        assert!(
-            !is_allowed_by_robots("https://example.com/private/secret", "example.com", &cache)
-                .await
-        );
-    }
-
-    #[test]
-    fn test_get_crawl_delay_returns_cached_value() {
-        let cache = new_robots_cache();
-        let rules = Arc::new(RobotsRules {
-            content: String::new(),
-            crawl_delay_secs: Some(7.5),
-        });
-        cache.insert("slow-site.com".to_string(), rules);
-
-        assert_eq!(get_crawl_delay("slow-site.com", &cache), Some(7.5));
-        assert_eq!(get_crawl_delay("unknown.com", &cache), None);
-    }
-
-    #[test]
-    fn test_robots_txt_empty_disallow_all() {
-        let robots_body = "User-agent: *\nDisallow: /\n";
-        let mut matcher = DefaultMatcher::default();
-        assert!(
-            !matcher.one_agent_allowed_by_robots(robots_body, "*", "https://example.com/anything"),
-            "Disallow: / should block everything"
-        );
-    }
-
-    #[test]
-    fn test_robots_txt_empty_permissive() {
-        let robots_body = "User-agent: *\n";
-        let mut matcher = DefaultMatcher::default();
-        assert!(
-            matcher.one_agent_allowed_by_robots(robots_body, "*", "https://example.com/anything"),
-            "Empty robots.txt (no Disallow) should allow everything"
-        );
-    }
-
-    // =========================================================================
-    // derive_filename_from_response tests
-    // =========================================================================
-
-    #[test]
-    fn test_derive_filename_pdf() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "application/pdf");
-        assert!(
-            result.ends_with(".pdf"),
-            "Expected .pdf extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_png() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "image/png");
-        assert!(
-            result.ends_with(".png"),
-            "Expected .png extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_jpg() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "image/jpeg");
-        assert!(
-            result.ends_with(".jpg"),
-            "Expected .jpg extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_gif() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "image/gif");
-        assert!(
-            result.ends_with(".gif"),
-            "Expected .gif extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_webp() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "image/webp");
-        assert!(
-            result.ends_with(".webp"),
-            "Expected .webp extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_mp3() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "audio/mpeg");
-        assert!(
-            result.ends_with(".mp3"),
-            "Expected .mp3 extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_mp4() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "video/mp4");
-        assert!(
-            result.ends_with(".mp4"),
-            "Expected .mp4 extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_zip() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "application/zip");
-        assert!(
-            result.ends_with(".zip"),
-            "Expected .zip extension, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_unknown() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "text/plain");
-        assert!(
-            result.ends_with(".bin"),
-            "Expected .bin extension for unknown type, got: {result}"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_from_url_path() {
-        let headers = wreq::header::HeaderMap::new();
-        let url = Url::parse("https://example.com/docs/report.pdf").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "application/octet-stream");
-        assert_eq!(
-            result, "report.pdf",
-            "URL basename should take priority over content-type"
-        );
-    }
-
-    #[test]
-    fn test_derive_filename_from_content_disposition() {
-        let mut headers = wreq::header::HeaderMap::new();
-        headers.insert(
-            wreq::header::CONTENT_DISPOSITION,
-            "attachment; filename=\"invoice.pdf\""
-                .parse()
-                .expect("valid header value"),
-        );
-        let url = Url::parse("https://example.com/download").expect("valid url");
-        let result = derive_filename_from_response(&headers, &url, "application/octet-stream");
-        assert_eq!(
-            result, "invoice.pdf",
-            "Content-Disposition should have highest priority"
-        );
-    }
-
-    // =========================================================================
-    // parse_content_disposition tests
-    // =========================================================================
-
-    #[test]
-    fn test_parse_content_disposition_filename() {
-        let result = parse_content_disposition("attachment; filename=\"report.pdf\"");
-        assert_eq!(result, Some("report.pdf".to_string()));
-    }
-
-    #[test]
-    fn test_parse_content_disposition_filename_unquoted() {
-        let result = parse_content_disposition("attachment; filename=report.pdf");
-        assert_eq!(result, Some("report.pdf".to_string()));
-    }
-
-    #[test]
-    fn test_parse_content_disposition_utf8() {
-        let result = parse_content_disposition("attachment; filename*=UTF-8''encoded.pdf");
-        assert_eq!(result, Some("encoded.pdf".to_string()));
-    }
-
-    #[test]
-    fn test_parse_content_disposition_empty() {
-        let result = parse_content_disposition("");
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn test_parse_content_disposition_no_filename() {
-        let result = parse_content_disposition("attachment");
-        assert_eq!(result, None);
     }
 }
