+++ src/cli/scrape_flow.rs
@@ -6,11 +6,11 @@ use tracing::{info, warn};
 use url::Url;
 
 use crate::application::crawl_options::CrawlOptions;
-use crate::application::crawler::discovery::{is_allowed_by_robots, new_robots_cache};
 use crate::application::export_factory;
 use crate::application::progress_types::{ScrapeError, ScrapeProgress, ScrapeStatus};
 use crate::application::scrape_single_url_for_tui;
 use crate::domain::ScrapedContent;
+use crate::infrastructure::crawler::robots_utils::{is_allowed_by_robots, new_robots_cache};
 use crate::infrastructure::export::state_store::StateStore;
 use crate::ScraperConfig;
 use crate::{HttpClient, HttpClientConfig};
